handle fallout of removing hadoop_extra_args from MRJobRunner
